web/storagenode: vue files renamed
