+++ web/storagenode/src/app/components/VBar.vue
