+++ web/storagenode/src/app/views/DashboardArea.vue
